# auslib/web/admin/swagger/api.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.
security:
  - basic: []
securityDefinitions:
  basic:
    type: basic
    description: HTTP basic access authentication

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
  usernameParam:
    name: username
    in: path
    description: unique identifier name of the user
    type: string
    maxLength: 100
    required: true
...
  csrfTokenParam:
    name: csrf_token
    in: query
    description: valid csrf token
    type: string
    required: true
  dataVersionParam:
    name: data_version
    in: query
    description: data version of object.
    type: integer
    format: int32
    minimum: 1
    required: true
...
  roleParam:
    name: role
    in: path
    description: user role to grant access to
    type: string
    maxLength: 50
    required: true
...
paths:
...
  /users/{username}/roles:
    get:
      summary: Returns list of user roles
      description: >
        Returns list of user roles as a json object. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-roles).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_get_roles
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id26"
        description: Returns list of user roles
      parameters:
        - $ref: '#/parameters/usernameParam'
      responses:
        '200':
          description: Successfully fetched the list of user roles.
          schema:
            type: object
            properties:
              roles:
                description: list of roles
                type: array
                uniqueItems: true
                minItems: 0
                items:
                  allOf:
                    - $ref: '#/definitions/DataVersionModel'
                    - $ref: '#/definitions/RoleModel'
                  required:
                    - data_version
                    - role
  /users/{username}/permissions:
    get:
      summary: Returns all of the details about the user
      description: >
        Returns all of the details about the user's permissions [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-permissions).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_permissions_get
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id21"
        description: Returns all of the details about the user's permissions
      parameters:
        - $ref: '#/parameters/usernameParam'
      responses:
        '200':
          description: Successfully fetched all permission details
          schema:
            $ref: "#/definitions/UserPermissionsGET"
          examples:
            application/json:
              rule:
                data_version: 2
                options:
                  actions: ['create', 'modify']
  /users/{username}/roles/{role}:
    put:
      summary: Grants the given username the given role
      description: >
        Grants the given username the given role. If the user already holds that role, this is a no-op. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-roles-role).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_role_put
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id27"
        description: Grants the given username the given role
      parameters:
        - $ref: '#/parameters/usernameParam'
        - $ref: '#/parameters/roleParam'
      responses:
        '200':
          $ref: '#/responses/modifiedDataVersion'
        '201':
          $ref: '#/responses/newObjectDataVersion'
    delete:
      summary: Revokes the given role from the given username
      description: >
        Revokes the given role from the given username. The 'data_version' parameter must be provided. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-roles-role).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_role_delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id28"
        description: Revokes the given role from the given username
      parameters:
        - $ref: '#/parameters/usernameParam'
        - $ref: '#/parameters/roleParam'
        - $ref: '#/parameters/csrfTokenParam'
        - $ref: '#/parameters/dataVersionParam'
      responses:
        '200':
          description: Successfully revoked the user role
        '404':
          $ref: '#/responses/resourceNotFound'
  /users/{username}/permissions/{permission}:
    get:
      summary: Returns the details of the named permission
      description: >
        Returns the details of the named permission for the username given in a JSON Object. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-permissions-permission).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_specific_permission_get
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id22"
        description: Returns the details of the named permission
      parameters:
        - $ref: '#/parameters/usernameParam'
        - $ref: '#/parameters/permissionParam'
      responses:
        '200':
          description: Successfully fetched the details of the given permission
          schema:
            $ref: "#/definitions/UserPermissionsModel"
          examples:
            application/json:
              data_version: 1
              options:
                actions:
                  - create
                  - modify
        '404':
          $ref: '#/responses/resourceNotFound'

    put:
      summary: Overwrites or creates the details of named permission for the username given.
      description: >
        Overwrites the details of named permission for the username given. If the permission does not exist, it is created. The following parameters are supported , data_version (required if the permission already exists) and options, a JSON string field [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-permissions-permission).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_specific_permission_put
      consumes:
        - application/json
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id23"
        description: Overwrites the options of named permission for the username given.If the permission does not exist, it is created first.
      parameters:
        - $ref: '#/parameters/usernameParam'
        - $ref: '#/parameters/permissionParam'
        - $ref: '#/parameters/userPermissionFormDataParam'
      responses:
        '200':
          $ref: '#/responses/updateExistingObject'
        '201':
          $ref: '#/responses/newObjectDataVersion'
        '400':
          $ref: '#/responses/invalidFormData'

    post:
      summary: Only overwrites the details of named permission for the username given.
      description: >
        Overwrites the details of named permission for the username given. The following parameters are supported , data_version(required) and options, a JSON string field [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-permissions-permission).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_specific_permission_post
      consumes:
        - application/json
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id23"
        description: Overwrites the options of named permission for the username given.
      parameters:
        - $ref: '#/parameters/usernameParam'
        - $ref: '#/parameters/permissionParam'
        - $ref: '#/parameters/userPermissionFormDataParam'
      responses:
        '200':
          $ref: '#/responses/updateExistingObject'
        '400':
          $ref: '#/responses/invalidFormData'
        '404':
          $ref: '#/responses/resourceNotFound'

    delete:
      summary: Deletes the named permission for the username given.
      description: >
        Deletes the named permission for the username given. data_version is required. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#users-username-permissions-permission).
      tags:
        - Users
      operationId: auslib.web.admin.views.mapper.user_specific_permission_delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id28"
        description: "Deletes the named permission for the username given. The 'data_version' parameter must be provided."
      parameters:
        - $ref: '#/parameters/usernameParam'
        - $ref: '#/parameters/permissionParam'
        - $ref: '#/parameters/csrfTokenParam'
        - $ref: '#/parameters/dataVersionParam'
      responses:
        '200':
          description: Successfully deleted the permission
        '400':
          $ref: '#/responses/invalidFormData'
        '404':
          $ref: '#/responses/resourceNotFound'
...
  /scheduled_changes/permissions:
    get:
      summary: Returns scheduled changes for permissions.
      description: >
        Endpoint to get Scheduled Changes for Permissions. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.scheduled_changes_permissions_get
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object"
        description: Returns scheduled changes for permissions.
      parameters:
        - $ref: '#/parameters/allSCParam'
      responses:
        '200':
          description: Returns JSON object of list of all scheduled changes for permissions plus total count
          schema:
            type: object
            required:
              - scheduled_changes
            allOf:
              - $ref: "#/definitions/CountModel"
              - type: object
                properties:
                  scheduled_changes:
                    description: array where each element is a scheduled change for a Permission.
                    type: array
                    minItems: 0
                    items:
                      type: object
                      allOf:
                        - $ref: '#/definitions/SCChangeType'
                        - $ref: '#/definitions/SCTime'
                        - $ref: '#/definitions/SCIdModel'
                        - $ref: '#/definitions/SCDataVersion'
                        - $ref: '#/definitions/SCSignoffsModel'
                        - $ref: '#/definitions/SCCommonFieldsModel'
                        - $ref: '#/definitions/DataVersionNullable'
                        - $ref: '#/definitions/OptionsModel'
                        - $ref: '#/definitions/PermissionsModel'
                        - properties:
                            original_row:
                              allOf:
                                - $ref: '#/definitions/DataVersionNullable'
                                - $ref: '#/definitions/PermissionsModel'
                                - $ref: '#/definitions/OptionsModel'
                      required:
                        - change_type
                        - data_version
                        - when
                        - permission
                        - username
    post:
      summary: create scheduled change for permissions.
      description: >
        Create scheduled change for a permission(existing or new). [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id34)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.scheduled_changes_permissions_post
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object"
        description: "create scheduled change for permissions."
      parameters:
        - name: sc_permission_body
          in: body
          description: request body for sc permission
          required: true
          schema:
            type: object
            allOf:
              - $ref: '#/definitions/CSRFModel'
              - $ref: '#/definitions/SCChangeType'
              - $ref: '#/definitions/SCTime'
              - $ref: '#/definitions/DataVersionNullable'
              - $ref: '#/definitions/PermissionsModel'
              - $ref: '#/definitions/OptionsJSON'
            required:
              - change_type
              - permission
              - username
              - when
      responses:
        '200':
          description: Returns sc_id of change
          schema:
            $ref: '#/definitions/SCIdModel'
        '400':
          $ref: '#/responses/invalidFormData'
  /scheduled_changes/permissions/{sc_id}:
    post:
      summary: Modifies existing scheduled change for a Permission.
      description: >
        Modifies existing scheduled change for a permission. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.scheduled_change_permissions_post
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id"
        description: "Modifies existing scheduled change for a Permission."
      parameters:
        - $ref: '#/parameters/sc_id_param'
        - name: sc_permission_body
          in: body
          description: request body for editing sc permission
          required: true
          schema:
            type: object
            allOf:
              - $ref: '#/definitions/CSRFModel'
              - $ref: '#/definitions/SCTime'
              - $ref: '#/definitions/SCDataVersion'
              - $ref: "#/definitions/DataVersionNullable"
              - $ref: '#/definitions/OptionsJSON'
              - $ref: '#/definitions/PermissionsModel'
      responses:
        '200':
          $ref: '#/responses/updateExistingObject'
        '400':
          $ref: '#/responses/invalidFormData'
        '404':
          $ref: '#/responses/resourceNotFound'

    delete:
      summary: Deletes the given Scheduled Change for the Permission.
      description: >
        Deletes the given Scheduled Change for the named Permission object. 'data_version' must be provided. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.scheduled_change_permissions_delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id"
        description: Deletes the given Scheduled Change for the Permission.
      parameters:
        - $ref: '#/parameters/sc_id_param'
        - $ref: '#/parameters/dataVersionParam'
        - $ref: '#/parameters/csrfTokenParam'
      responses:
        '200':
          description: Successfully deleted the given scheduled change.
        '400':
          $ref: '#/responses/invalidFormData'
        '404':
          $ref: '#/responses/resourceNotFound'
...
  /scheduled_changes/permissions/{sc_id}/enact:
    post:
      summary: Execute the scheduled change for the permission.
      description: >
        Execute the scheduled change for the permission identified by the sc_id. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-enact)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.enact_scheduled_change_permissions_post
      consumes: []
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-enact"
        description: "Execute the scheduled change for the permission."
      parameters:
        - $ref: '#/parameters/sc_id_param'
      responses:
        '200':
          description: Successfully enacted the scheduled change
...
  /scheduled_changes/permissions/{sc_id}/signoffs:
    post:
      summary: Signs off on the given Scheduled Change for Permission.
      description: >
        Signs off on the given Scheduled Change for Permission. 'role' must be provided in the request body.[Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-signoffs)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.scheduled_change_permissions_signoffs_post
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-signoffs"
        description: "Signs off on the given Scheduled Change for Permission."
      parameters:
        - $ref: '#/parameters/sc_id_param'
        - $ref: '#/parameters/sc_signoff_post_body_param'
      responses:
        '200':
          description: Successfully signed off on the given scheduled change for the named object.
        '400':
          $ref: '#/responses/invalidFormData'

    delete:
      summary: Removes an existing Signoff from the Scheduled Change for the Permission.
      description: >
        Removes an existing Signoff from the Scheduled Change for the Permission. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-signoffs)
      tags:
        - Scheduled Changes
      operationId: auslib.web.admin.views.mapper.scheduled_change_permissions_signoffs_delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#scheduled-changes-object-sc-id-signoffs"
        description: Removes an existing Signoff from the Scheduled Change for the Permission.
      parameters:
        - $ref: '#/parameters/sc_id_param'
        - $ref: '#/parameters/usernameQueryParam'
        - $ref: '#/parameters/csrfTokenParam'
      responses:
        '200':
          description: Successfully deleted the signoff on the given scheduled change.
        '404':
          $ref: '#/responses/resourceNotFound'
...
definitions:
...
  RoleModel:
    title: User Role
    description: "model to re-use 'role' field."
    type: object
    properties:
      role:
        description: roles are used when signing off on Scheduled Changes
        type: string
        minLength: 1
        maxLength: 50
        example: releng
...